remove compile warning of logError

+++ common/http_func.c
@@ -69,8 +69,8 @@ int get_url_content(const char *url, const int timeout, int *http_status, \
 	if (domain_len >= sizeof(domain_name))
 	{
 		logError("file: "__FILE__", line: %d, " \
-			"domain is too large, exceed %ld.", \
-			__LINE__, sizeof(domain_name));
+			"domain is too large, exceed %d.", \
+			__LINE__, (int)sizeof(domain_name));
 		return EINVAL;
 	}
